feat: variable suggestions popover
