+++ packages/core/src/editor/components/ui/input-autocomplete.tsx
@@ -1,13 +1,13 @@
+import { VariableSuggestionsPopoverRef } from '@/editor/nodes/variable/variable-suggestions-popover';
 import {
-  VariablePopover,
-  VariablePopoverRef,
-} from '@/editor/nodes/variable/variable-popover';
-import { useMailyContext } from '@/editor/provider';
+  DEFAULT_VARIABLE_SUGGESTION_POPOVER,
+  useMailyContext,
+} from '@/editor/provider';
 import { cn } from '@/editor/utils/classname';
 import { AUTOCOMPLETE_PASSWORD_MANAGERS_OFF } from '@/editor/utils/constants';
 import { useOutsideClick } from '@/editor/utils/use-outside-click';
 import { CornerDownLeft } from 'lucide-react';
-import { forwardRef, HTMLAttributes, useMemo, useState, useRef } from 'react';
+import { forwardRef, HTMLAttributes, useRef } from 'react';
 
 type InputAutocompleteProps = HTMLAttributes<HTMLInputElement> & {
   value: string;
@@ -37,7 +37,11 @@ export const InputAutocomplete = forwardRef<
   } = props;
 
   const containerRef = useRef<HTMLDivElement>(null);
-  const popoverRef = useRef<VariablePopoverRef>(null);
+  const popoverRef = useRef<VariableSuggestionsPopoverRef>(null);
+  const {
+    variableSuggestionPopover:
+      VariableSuggestionPopoverComponent = DEFAULT_VARIABLE_SUGGESTION_POPOVER,
+  } = useMailyContext();
 
   useOutsideClick(containerRef, () => {
     onOutsideClick?.();
@@ -88,7 +92,7 @@ export const InputAutocomplete = forwardRef<
 
       {isTriggeringVariable && (
         <div className="mly-absolute mly-left-0 mly-top-8">
-          <VariablePopover
+          <DEFAULT_VARIABLE_SUGGESTION_POPOVER
             items={autoCompleteOptions.map((option) => {
               return {
                 name: option,
